extract - external:   - Default detector to 'external' when importing alignments   - Handle different frame origin co-ordinates
