+++ plugins/extract/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+""" Package for Faceswap's extraction pipeline """
+from .extract_media import ExtractMedia
+from .pipeline import Extractor
