+++ plugins/extract/align/_base/aligner.py
@@ -4,7 +4,7 @@
 All Aligner Plugins should inherit from this class.
 See the override methods for which methods are required.
 
-The plugin will receive a :class:`~plugins.extract.pipeline.ExtractMedia` object.
+The plugin will receive a :class:`~plugins.extract.extract_media.ExtractMedia` object.
 
 For each source item, the plugin must pass a dict to finalize containing:
 
@@ -26,7 +26,8 @@ from tensorflow.python.framework import errors_impl as tf_errors  # pylint:disab
 
 from lib.align import LandmarkType
 from lib.utils import FaceswapError
-from plugins.extract._base import BatchType, Extractor, ExtractMedia, ExtractorBatch
+from plugins.extract import ExtractMedia
+from plugins.extract._base import BatchType, ExtractorBatch, Extractor
 from .processing import AlignedFilter, ReAlign
 
 if T.TYPE_CHECKING:
@@ -242,8 +243,8 @@ class Aligner(Extractor):  # pylint:disable=abstract-method
         Items are returned from the ``queue`` in batches of
         :attr:`~plugins.extract._base.Extractor.batchsize`
 
-        Items are received as :class:`~plugins.extract.pipeline.ExtractMedia` objects and converted
-        to ``dict`` for internal processing.
+        Items are received as :class:`~plugins.extract.extract_media.ExtractMedia` objects and
+        converted to ``dict`` for internal processing.
 
         To ensure consistent batch sizes for aligner the items are split into separate items for
         each :class:`~lib.align.DetectedFace` object.
@@ -352,7 +353,7 @@ class Aligner(Extractor):  # pylint:disable=abstract-method
 
         Yields
         ------
-        :class:`~plugins.extract.pipeline.ExtractMedia`
+        :class:`~plugins.extract.extract_media.ExtractMedia`
             The :attr:`DetectedFaces` list will be populated for this class with the bounding boxes
             and landmarks for the detected faces found in the frame.
         """
