+++ plugins/extract/detect/_base.py
@@ -4,7 +4,7 @@
 All Detector Plugins should inherit from this class.
 See the override methods for which methods are required.
 
-The plugin will receive a :class:`~plugins.extract.pipeline.ExtractMedia` object.
+The plugin will receive a :class:`~plugins.extract.extract_media.ExtractMedia` object.
 
 For each source frame, the plugin must pass a dict to finalize containing:
 
@@ -30,7 +30,7 @@ from lib.align import DetectedFace
 from lib.utils import FaceswapError
 
 from plugins.extract._base import BatchType, Extractor, ExtractorBatch
-from plugins.extract.pipeline import ExtractMedia
+from plugins.extract import ExtractMedia
 
 if T.TYPE_CHECKING:
     from collections.abc import Generator
@@ -132,8 +132,8 @@ class Detector(Extractor):  # pylint:disable=abstract-method
     def get_batch(self, queue: Queue) -> tuple[bool, DetectorBatch]:
         """ Get items for inputting to the detector plugin in batches
 
-        Items are received as :class:`~plugins.extract.pipeline.ExtractMedia` objects and converted
-        to ``dict`` for internal processing.
+        Items are received as :class:`~plugins.extract.extract_media.ExtractMedia` objects and
+        converted to ``dict`` for internal processing.
 
         Items are returned from the ``queue`` in batches of
         :attr:`~plugins.extract._base.Extractor.batchsize`
@@ -208,7 +208,7 @@ class Detector(Extractor):  # pylint:disable=abstract-method
 
         Yields
         ------
-        :class:`~plugins.extract.pipeline.ExtractMedia`
+        :class:`~plugins.extract.extract_media.ExtractMedia`
             The :attr:`DetectedFaces` list will be populated for this class with the bounding boxes
             for the detected faces found in the frame.
         """
@@ -351,7 +351,7 @@ class Detector(Extractor):  # pylint:disable=abstract-method
 
         Parameters
         ----------
-        item: :class:`plugins.extract.pipeline.ExtractMedia`
+        item: :class:`~plugins.extract.extract_media.ExtractMedia`
             The input item from the pipeline
 
         Returns
