+++ scripts/convert.py
@@ -23,7 +23,7 @@ from lib.image import read_image_meta_batch, ImagesLoader
 from lib.multithreading import MultiThread, total_cpus
 from lib.queue_manager import queue_manager
 from lib.utils import FaceswapError, get_folder, get_image_paths, handle_deprecated_cliopts
-from plugins.extract.pipeline import Extractor, ExtractMedia
+from plugins.extract import ExtractMedia, Extractor
 from plugins.plugin_loader import PluginLoader
 
 if T.TYPE_CHECKING:
@@ -44,7 +44,7 @@ class ConvertItem:
 
     Parameters
     ----------
-    input: :class:`~plugins.extract.pipeline.ExtractMedia`
+    input: :class:`~plugins.extract.extract_media.ExtractMedia`
         The ExtractMedia object holding the :attr:`filename`, :attr:`image` and attr:`list` of
         :class:`~lib.align.DetectedFace` objects loaded from disk
     feed_faces: list, Optional
@@ -1094,7 +1094,7 @@ class Predict():
         logger.trace("Queued out batch. Batchsize: %s", len(batch))  # type:ignore
 
 
-class OptionalActions():
+class OptionalActions():  # pylint:disable=too-few-public-methods
     """ Process specific optional actions for Convert.
 
     Currently only handles skip faces. This class should probably be (re)moved.
