+++ scripts/fsmedia.py
@@ -25,7 +25,7 @@ if T.TYPE_CHECKING:
     from collections.abc import Generator
     from argparse import Namespace
     from lib.align import AlignedFace
-    from plugins.extract.pipeline import ExtractMedia
+    from plugins.extract import ExtractMedia
 
 logger = logging.getLogger(__name__)
 
@@ -414,14 +414,15 @@ class PostProcess():
 
         Parameters
         ----------
-        extract_media: :class:`~plugins.extract.pipeline.ExtractMedia`
-            The :class:`~plugins.extract.pipeline.ExtractMedia` object to perform the
+        extract_media: :class:`~plugins.extract.extract_media.ExtractMedia`
+            The :class:`~plugins.extract.extract_media.ExtractMedia` object to perform the
             action on.
 
         Returns
         -------
-        :class:`~plugins.extract.pipeline.ExtractMedia`
-            The original :class:`~plugins.extract.pipeline.ExtractMedia` with any actions applied
+        :class:`~plugins.extract.extract_media.ExtractMedia`
+            The original :class:`~plugins.extract.extract_media.ExtractMedia` with any actions
+            applied
         """
         for action in self._actions:
             logger.debug("Performing postprocess action: '%s'", action.__class__.__name__)
@@ -458,8 +459,8 @@ class PostProcessAction():
 
         Parameters
         ----------
-        extract_media: :class:`~plugins.extract.pipeline.ExtractMedia`
-            The :class:`~plugins.extract.pipeline.ExtractMedia` object to perform the
+        extract_media: :class:`~plugins.extract.extract_media.ExtractMedia`
+            The :class:`~plugins.extract.extract_media.ExtractMedia` object to perform the
             action on.
         """
         raise NotImplementedError
@@ -578,9 +579,9 @@ class DebugLandmarks(PostProcessAction):
 
         Parameters
         ----------
-        extract_media: :class:`~plugins.extract.pipeline.ExtractMedia`
-            The :class:`~plugins.extract.pipeline.ExtractMedia` object that contains the faces to
-            draw the landmarks on to
+        extract_media: :class:`~plugins.extract.extract_media.ExtractMedia`
+            The :class:`~plugins.extract.extract_media.ExtractMedia` object that contains the faces
+            to draw the landmarks on to
         """
         frame = os.path.splitext(os.path.basename(extract_media.filename))[0]
         for idx, face in enumerate(extract_media.detected_faces):
